+++ packages/iov-encoding/types/decimal.d.ts
@@ -8,7 +8,19 @@ export declare class Decimal {
   static fromAtomics(atomics: string, fractionalDigits: number): Decimal;
   private static verifyFractionalDigits;
   readonly atomics: string;
+  readonly fractionalDigits: number;
   private readonly data;
   private constructor();
   toString(): string;
+  /**
+   * Returns an approximation as a float type. Only use this if no
+   * exact calculation is required.
+   */
+  toFloatApproximation(): number;
+  /**
+   * a.plus(b) returns a+b.
+   *
+   * Both values need to have the same fractional digits.
+   */
+  plus(b: Decimal): Decimal;
 }