+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   multisignature contracts
 - @iov/encoding: Add missing integer check on input bytes in
   `Uint32.fromBigEndianBytes` and `Uint64.fromBytesBigEndian`
+- @iov/encoding: Add `Decimal.fractionalDigits`, `.plus` and
+  `.toFloatApproximation`.
 
 ## 0.17.7